requirements.txt
aioredis==1.2.0
altgraph==0.16.1
amqp==1.4.9
ansible==2.4.2.0
anyjson==0.3.3
asgi-redis==1.4.3
asgiref==2.3.2
asn1crypto==0.24.0
async==0.6.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
backcall==0.1.0
bcrypt==3.1.4
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2018.8.24
cffi==1.11.5
channels==2.1.5
channels-redis==2.3.1
chardet==3.0.4
constantly==15.1.0
crypto==1.4.1
cryptography==2.3.1
daemon==1.2
daphne==2.2.2
decorator==4.3.0
Django==2.0.8
django-filter==2.0.0
django-mdeditor==0.1.6
django-suit==2.0a1
djangorestframework==3.8.2
dnspython==1.15.0
eventlet==0.24.1
future==0.17.1
greenlet==0.4.15
hiredis==0.2.0
httplib2==0.11.3
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
Jinja2==2.10
kombu==3.0.37
macholib==1.11
Markdown==2.6.11
MarkupSafe==1.0
monotonic==1.5
msgpack==0.5.6
msgpack-python==0.5.6
Naked==0.1.31
ntlm-auth==1.2.0
paramiko==2.4.1
parso==0.3.1
pefile==2018.8.8
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
ply==3.11
prompt-toolkit==1.0.15
psutil==5.4.7
ptyprocess==0.6.0
pyasn1==0.4.4
pycparser==2.18
pycrypto==2.6.1
pycryptodomex==3.6.6
Pygments==2.2.0
PyHamcrest==1.9.0
pyinstall==0.1.4
PyInstaller==3.4
PyMySQL==0.9.2
PyNaCl==1.2.1
pysmi==0.3.1
pysnmp==4.4.5
python-nmap==0.6.1
pytz==2018.5
pyvmomi==6.7.0.2018.9
pywinrm==0.3.0
PyYAML==3.13
redis==2.10.6
requests==2.19.1
requests-ntlm==1.1.0
shellescape==3.4.1
simplegeneric==0.8.1
six==1.11.0
tornado==5.1.1
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
urllib3==1.23
utils==0.9.0
vine==1.1.4
wcwidth==0.1.7
webssh==0.7.1
xmltodict==0.11.0
zope.interface==4.6.0
